Introduce new unified sd card layout.
The new layout uses two partitions (like the old "advanced layout"),
but keeps the main system image on the first (FAT32) partition
like the old "simple layout". The second partition is formatted as ext4
and holds user data (in /home) as well as writeable overlays for /etc
(for individual configuration data, e.g. passwords, ssh keys, ...) and
(optionally) for the file system root.

The choice between both modes (read-only or writeable /) can be made
by placing a file "rootfs.conf" on the first partition. If this file
exists and has the content "rw", the TXT will boot up with a writeable
overlay for "/". If "rootfs.conf" is not present or has a different
content than "rw", the file system root will be read-only.

In both modes, "/etc" will be placed on a writeably overlay, and "/home"
will be mounted from the writeable second partition.

The second partition must contain the following directories:
* home/ftc - owned by "ftc" (uid 1001, gid 1001), mode u=rwx,go=rx
* home/ftc/apps - owner and permissions same as /home/ftc
* home/ROBOPro - owned by "ROBOPro" (uid 1000, gid 1000), mode u=rwx,go=rx
* home/ROBOPro/ROBOProFiles - owner and permissions same as /home/ROBOPro
* overlay - owned by "root", permissions u=rwx,go-rwx
* overlay/.workdir/etc - owned by root, permissions u=rwx,go=rx
* overlay/etc - owned by root, permissions u=rwx,go=rx

Additional directories for the root overlay (overlay/rootfs and
overlay/.workdir/rootfs) will be autmatically created as necessary.