Enhance version handling in the store app, allow app updates from version specific branch
The store app now interprets the 'firmware' entry in app manifests as a version
specification, and shows only apps where this spec matches the firmware version
from /etc/fw-ver.txt, and the store app now additionally looks for apps in a
branch with a name matching the firmware version /etc/fw-ver.txt.

If both the master and the firmware specific branch contain a compatible version
of the same app, the store chooses the newer version as indicated by the 'version'
field in the app manifest. If both apps have the same version, the store uses the
variant from the 'master' branch.