Workaround for segfault while installing apps
After the base system update to buildroot 2017.02.1, the store app segfaults while executing z.read("manifest") (zipfile read) in store.py line 280 when the store is run as a launcher plugin. If the store is run standalone (i.e. in a separate python process), everything is fine.

Until this issue is resolved and fixed, revert to running the store app standalone as a workaround.